InvalidOperationException message changed

+++ WebApi/Applications/AuthorOperations/Command/DeleteAuthor/DeleteAuthorCommand.cs
@@ -16,13 +16,15 @@ namespace WebApi.Applications.AuthorOperations.Command.DeleteAuthor
         public int AuthorId { get; set; }
         public void Handle()
         {
-            
+
             var author = _context.Authors.SingleOrDefault(x => x.Id == AuthorId);
             if (author is null)
-                throw new InvalidOperationException(AuthorId + " numaralı silinecek yazar bulunamadı");
+                throw new InvalidOperationException("Silinecek yazar bulunamadı");
+
 
             var books = _context.Books.Where(x => x.AuthorId == AuthorId).ToList();
-            _context.Books.RemoveRange(books);
+            if (books.Count > 0)
+                _context.Books.RemoveRange(books);
 
             _context.Authors.Remove(author);
             _context.SaveChanges();
